pyproject.toml
[project]
name = "ai-health-and-wellness-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.1",
    "fpdf>=1.7.2",
    "google-generativeai>=0.8.5",
    "load-dotenv",
    "openai-agent",
    "streamlit",
]
